chore: cleanup and reorg of code
