--- packages/common/misc/asn1/types/ecdsaPublicKey.ts
@@ -1,10 +1,0 @@
-import { AsnProp, AsnPropTypes } from '@peculiar/asn1-schema';
-import { ECDSAPublicKeyDetais } from './ecdsaPublicKeyDetails';
-
-export class ECDSAPublicKey {
-  @AsnProp({ type: ECDSAPublicKeyDetais })
-  public publicKeyDetails = new ECDSAPublicKeyDetais();
-
-  @AsnProp({ type: AsnPropTypes.BitString })
-  public publicKey = new ArrayBuffer(0);
-}
