--- packages/common/misc/asn1/types/ecdsaPublicKeyDetails.ts
@@ -1,13 +1,0 @@
-import { AsnProp, AsnPropTypes } from '@peculiar/asn1-schema';
-
-export class ECDSAPublicKeyDetais {
-  @AsnProp({ type: AsnPropTypes.ObjectIdentifier })
-  public type = '';
-
-  @AsnProp({ type: AsnPropTypes.ObjectIdentifier })
-  public namedCurve = '';
-
-  constructor(params: Partial<ECDSAPublicKeyDetais> = {}) {
-    Object.assign(this, params);
-  }
-}
