+++ packages/contracts/package.json
@@ -11,6 +11,7 @@
     "hh:clean": "hardhat clean"
   },
   "devDependencies": {
+    "@giano/common": "workspace:",
     "@nomicfoundation/hardhat-chai-matchers": "^2.0.2",
     "@nomicfoundation/hardhat-ethers": "^3.0.5",
     "@nomicfoundation/hardhat-ignition": "^0.15.7",
