+++ packages/contracts/test/Account.ts
@@ -1,12 +1,10 @@
+import { encodeChallenge, hexToUint8Array } from '@giano/common';
+import { loadFixture } from '@nomicfoundation/hardhat-toolbox/network-helpers';
 import { expect } from 'chai';
 import { ethers, ignition } from 'hardhat';
-import { Account } from '../typechain-types';
-import { createKeypair } from './utils';
-import { ECDSASigValue } from '@peculiar/asn1-ecc';
-import { AsnParser } from '@peculiar/asn1-schema';
-import crypto from 'crypto';
 import GianoModule from '../ignition/modules/Giano';
-import { loadFixture } from '@nomicfoundation/hardhat-toolbox/network-helpers';
+import type { Account } from '../typechain-types';
+import { createKeypair, signWebAuthnChallenge } from './utils';
 
 describe('Account Contract', () => {
   let account: Account;
@@ -50,7 +48,7 @@ describe('Account Contract', () => {
 
       // Generate a valid signature
       const challenge = await account.getChallenge();
-      const signature = await signChallenge(keypair.keyPair.privateKey, hexToUint8Array(challenge));
+      const signature = signWebAuthnChallenge(keypair.keyPair.privateKey, hexToUint8Array(challenge));
 
       // Execute transfer via Account contract
       await expect(
@@ -76,7 +74,7 @@ describe('Account Contract', () => {
       const transferData = genericERC20.interface.encodeFunctionData('transfer', [recipient, amount]);
 
       // Generate an invalid signature (e.g., random bytes)
-      const invalidSignature = crypto.randomBytes(130); // 65 bytes for ECDSA
+      const invalidSignature = encodeChallenge(signWebAuthnChallenge(keypair.keyPair.privateKey, hexToUint8Array('0xDEADBEEF')));
 
       // Attempt to execute transfer via Account contract
       await expect(
@@ -86,75 +84,7 @@ describe('Account Contract', () => {
           data: transferData,
           signature: invalidSignature,
         }),
-      ).to.be.revertedWith('InvalidSignature');
+      ).to.be.revertedWithCustomError(account, 'InvalidSignature');
     });
   });
 });
-
-function hexToUint8Array(hex: string) {
-  if (hex.startsWith('0x')) {
-    hex = hex.slice(2);
-  }
-  return new Uint8Array(hex.match(/.{1,2}/g)!.map((byte) => parseInt(byte, 16)));
-}
-
-function uint8ArrayToUint256(array: ArrayBuffer) {
-  const hex = Array.from(new Uint8Array(array))
-    .map((b) => b.toString(16).padStart(2, '0'))
-    .join('');
-  return BigInt('0x' + hex);
-}
-
-function signChallenge(privateKey: crypto.KeyObject, challenge: Uint8Array) {
-  // Step 2: Prepare clientDataJSON
-  const clientData = {
-    type: 'webauthn.get',
-    challenge: Buffer.from(challenge).toString('base64url'),
-    origin: 'https://localhost:3000',
-    crossOrigin: false,
-  };
-  const clientDataJSON = Buffer.from(JSON.stringify(clientData));
-
-  // Step 3: Hash the clientDataJSON
-  const clientDataHash = crypto.createHash('sha256').update(clientDataJSON).digest();
-
-  // Step 4: Prepare authenticatorData
-  const rpIdHash = crypto.createHash('sha256').update('localhost').digest();
-  const flags = Buffer.from([0x01]); // User Present flag
-  const signCount = Buffer.alloc(4); // 32-bit signature counter
-  const authenticatorData = Buffer.concat([rpIdHash, flags, signCount]);
-
-  // Step 5: Concatenate authenticatorData and clientDataHash
-  const dataToSign = Buffer.concat([authenticatorData, clientDataHash]);
-
-  // Step 6: Sign the concatenated data
-  const signature = crypto.createSign('SHA256').update(dataToSign).sign(privateKey);
-
-  // Step 7: Assemble the response
-  return {
-    clientDataJSON: clientDataJSON,
-    authenticatorData: authenticatorData,
-    signature: signature,
-  };
-}
-
-function encodeChallenge(assertionResponse: any) {
-  const decodedClientDataJson = new TextDecoder().decode(assertionResponse.clientDataJSON);
-  const responseTypeLocation = decodedClientDataJson.indexOf('"type":');
-  const challengeLocation = decodedClientDataJson.indexOf('"challenge":');
-  const parsedSignature = AsnParser.parse(assertionResponse.signature, ECDSASigValue);
-
-  return ethers.AbiCoder.defaultAbiCoder().encode(
-    ['tuple(bytes authenticatorData, string clientDataJSON, uint256 challengeLocation, uint256 responseTypeLocation, uint256 r, uint256 s)'],
-    [
-      [
-        new Uint8Array(assertionResponse.authenticatorData),
-        decodedClientDataJson,
-        challengeLocation,
-        responseTypeLocation,
-        uint8ArrayToUint256(parsedSignature.r),
-        uint8ArrayToUint256(parsedSignature.s),
-      ],
-    ],
-  );
-}
