+++ services/web/package.json
@@ -14,6 +14,7 @@
     "web:serve": "serve -s dist -p 80 --config ../serve.json"
   },
   "dependencies": {
+    "@giano/client": "workspace:",
     "@giano/common": "workspace:",
     "@giano/contracts": "workspace:",
     "@peculiar/asn1-ecc": "^2.3.14",
